코드 최적화 - JPA n+1 이슈, JPA delete 이슈 등

+++ src/main/java/com/simple/app/controller/PostController.java
@@ -54,8 +54,13 @@ public class PostController {
         return Response.success();
     }
 
+//    @GetMapping("/{postId}/likes")
+//    public Response<Integer> likeCount(@PathVariable Integer postId, Authentication authentication) {
+//        return Response.success(postService.likeCount(postId));
+//    }
+
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(@PathVariable Integer postId, Authentication authentication) {
+    public Response<Long> likeCount(@PathVariable Integer postId, Authentication authentication) {
         return Response.success(postService.likeCount(postId));
     }
